Renames coordinator and nests message.
